requirements.txt
GitPython==3.1.2
PyYAML==5.3.1
requests==2.23.0
urllib3==1.25.9
